Moved ResourcePtr from internal namespace into public namespace, so it can be re-used by components outside libunity-api.
Removed dependency on boost from ResourcePtr.h. Functionally, there is no difference, it's just that the comparison operators
are now no longer guarded by enable_if (which depended on boost type traits). This will cause a somewhat less comprehensible
error message if someone calls a comparison operator on a ResourcePtr that is instantiated with a class that doesn't support
comparison, but otherwise has the exact same semantics. (Re-implementing the boost type traits was not an option because
they are way too complex to make that feasible.)